Support 32bits architectures.
- Introduce internal `ASize` and `ARegion` (A stands for Addressable or
Architecture independent) size or region. This is mainly a wrapper around
usize where `Size` is around u64.
- Correctly use `ASize` when we are loading the data in memory.

This may lead to some performance drawback or panic in 32 bits as we may
not be able to load huge data in memory.

+++ src/creator/container_pack.rs
@@ -27,7 +27,7 @@ impl ContainerPackCreator<NamedFile> {
 impl<F: PackRecipient> ContainerPackCreator<F> {
     pub fn from_file(mut file: Box<F>, free_data: PackFreeData) -> Result<Self> {
         file.seek(SeekFrom::Start(
-            PackHeader::BLOCK_SIZE as u64 + ContainerPackHeader::BLOCK_SIZE as u64,
+            (PackHeader::BLOCK_SIZE + ContainerPackHeader::BLOCK_SIZE) as u64,
         ))?;
         Ok(ContainerPackCreator {
             packs: vec![],
